requirements.txt
azureml==0.2.7
azureml-core==1.43.0
azure-storage-blob==12.12.0
mlflow==1.27.0
torch==1.12.0
torchvision==0.13.0
sklearn==0.0
albumentations==1.2.0
opencv-python==4.6.0.66
azureml-mlflow==1.43.0
azureml-dataset-runtime==1.43.0